do not push the unsigned artifact to github
slightly change the artifact name to lowercase gvim
Appveyor documentation mentions that the artifact key at the deploy
provider has to match either the path or the name. Make the name all
lowercase and then change the artifact key for the github provider to
only match those artifacts that start with ^gvim (that should prevent
the `unsigned-gvim_*` artifact from being uploaded to github
(hopefully).