Add Win AARCH64 feature to appveyor.bat script
Unfortunately, Ruby and Racket doesn't seem to be build successfully, so
those are skipped.

For the gettext libraries we will use vcpkg to build the Windows ARM64
versions of libintl-8.dll and libiconv-2.dll and package those

For libsodium we will compile our own AARCH64 version of the
libsodium.dll

Skip packaging the pdb zip file.

Signed-off-by: Christian Brabandt <cb@256bit.org>